+++ components/common/RoundButton.tsx
@@ -1,11 +1,7 @@
 import React from 'react';
 
-const RoundButton = ({ children }) => {
-    return (
-        <div>
-            {children}
-        </div>
-    );
+export const RoundButton = ({ children }) => {
+  return <div>{children}</div>;
 };
 
-export default RoundButton;
+// export default RoundButton;
