+++ components/common/index.tsx
@@ -1,0 +1,3 @@
+export * from './Header';
+export * from './Responsive';
+export * from './RoundButton';
