+++ components/index.ts
@@ -1,0 +1,2 @@
+export * from './auth/index';
+export * from './common/index';
