+++ package.json
@@ -22,6 +22,7 @@
   "devDependencies": {
     "@types/node": "^16.11.10",
     "@types/react": "^17.0.37",
+    "@types/styled-components": "^5.1.15",
     "eslint": "7.32.0",
     "eslint-config-next": "12.0.4",
     "typescript": "^4.5.2"
