+++ styles/globals.css
@@ -7,14 +7,16 @@ html,
 body {
   padding: 0;
   margin: 0;
-  font-family: 'Noto Sans KR', -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
-    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
+  font-family: 'Noto Sans KR', -apple-system, BlinkMacSystemFont, Segoe UI,
+    Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue,
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
 
-    font-weight: 400;
-    font-size: 16pt; /* 본문 16pt */
-    color: #333333;
+  font-weight: 400;
+  font-size: 1rem;
+  font-size: 16pt; /* 본문 16pt */
+  color: #333333;
 }
 
 a {
